// Client/src/app/subscription.service.ts
export interface PushSubscription {
    endPoint?: string | undefined;
    auth?: string | undefined;
    p256Dh?: string | undefined;
}